Update parts of module for Terraform 0.13
Why these changes are being introduced:
To prep for the upgrade of this module's actual Terraform code to be
compatible with Terraform 0.13, we need to update the automated
validation.

How this addresses that need:
* TravisCI configuration is deleted
* Makefile is deleted (it pointed to TravisCI)
* Pull Request Template added
* new tf-validate GitHub Action workflow added

Side effects of this change:
None.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/IN-324
* https://mitlibraries.atlassian.net/browse/IN-325
* https://mitlibraries.atlassian.net/browse/IN-326

On branch update-0.13
Changes to be committed:
	new file:   .github/PULL_REQUEST_TEMPLATE.md
	new file:   .github/workflows/tf-validate.yml
	deleted:    .travis.yml
	deleted:    Makefile